+++ app/src/main/java/com/aitor/cebancpizza/InformacionBebidas.java
@@ -9,12 +9,12 @@ import java.io.Serializable;
 public class InformacionBebidas implements Serializable{
     private String tipo;
     private int cantidad;
-    private double total;
+    private float total;
 
     public void setTipo(String tipo){this.tipo=tipo;}
     public void setCantidad(int cantidad){this.cantidad=cantidad;}
-    public void setTotal(double precio){this.total=precio;}
+    public void setTotal(float precio){this.total=precio;}
     public String getTipo(){return tipo;}
     public int getCantidad(){return cantidad;}
-    public double getTotal(){return total;}
+    public float getTotal(){return total;}
 }
